Implement good default behavior
After uploading the sketch to the Arduino you should be able to just plug
everything in and it should work well for a simple system.

Assume that the time is 10:00 and run the pump every day at 10:00 and
20:00 for at most one minute. Spray every 10 min for 2 sec each.

Refactor some of the horrible code. Next step will probably be to add unit
and integration tests.

Signed-off-by: Robin hagman <Robin hagman>