Update README.md
Remove dashes from the first header since it looks nicer on the github pages.
Add an image of the basic kit version in the window, the image is hosted on Dropbox. Better than nothing.
Add link to the shared Dropbox folder with a lot if different images of the project.