Use PlatformIO and Travis
PlatformIO seems like a good tool to use since I easily can build for
Arduino UNO and D1 mini. As well as other platforms in the future. Nice
support for it in the Atom IDE that I have started to like more and more.

Also, implement the use of Travis.

Signed-off-by: Robin Hagman <robin.hagman@gmail.com>