Update default settings
We now run the system with the same schedule and settings all the time.
When the microcontroller starts it assumes that the time is 20:00 and it
runs the pump. Then every 20min it sprays the roots for 2s.

Also make some style changes to the code and clean up a k = k++ wtf...

Signed-off-by: Robin Hagman <robin.hagman@gmail.com>